add private chat and update file names

+++ controllers/chatroom_controller.go
@@ -12,7 +12,7 @@ import (
 )
 
 func CreateGroup(c *gin.Context) {
-	var request dto.CreateGroupDto
+	var request dto.CreateChatDto
 
 	if err := c.ShouldBindJSON(&request); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": err.Error()})
@@ -50,7 +50,8 @@ func CreateGroup(c *gin.Context) {
 	}
 
 	if len(existingUsers) > 0 {
-		group := models.Group{
+		group := models.ChatRoom{
+			Type: 	     "group",
 			Name:        request.Name,
 			Description: request.Description,
 			Members:     existingUsers,
@@ -67,9 +68,9 @@ func CreateGroup(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"success": true, "message": "Group created successfully"})
 }
 
-func GetAllGroups(c *gin.Context) {
+func GetAllChatrooms(c *gin.Context) {
 	userId := c.Param("id")
-	var groups []models.Group
+	var groups []models.ChatRoom
 
 	if err := database.Db.Preload("Members").Where("members.id = ?", userId).Find(&groups).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
@@ -81,19 +82,19 @@ func GetAllGroups(c *gin.Context) {
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"success": true, "groups": groups})
+	c.JSON(http.StatusOK, gin.H{"success": true, "data": groups})
 }
 
-func GetGroup(c *gin.Context) {
+func GetChatroom(c *gin.Context) {
 	groupId := c.Param("id")
-	var group models.Group
+	var group models.ChatRoom
 
 	if err := database.Db.Preload("Members").Where("id = ?", groupId).First(&group).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"success": true, "group": group})
+	c.JSON(http.StatusOK, gin.H{"success": true, "data": group})
 }
 
 func AddMembers(c *gin.Context) {
@@ -110,13 +111,18 @@ func AddMembers(c *gin.Context) {
 		return
 	}
 
-	var group models.Group
+	var group models.ChatRoom
 	if err := database.Db.Where("id = ?", groupId).First(&group).Error; err != nil {
 		if err == gorm.ErrRecordNotFound {
 			c.JSON(http.StatusNotFound, gin.H{"success": false, "message": "Group not found"})
 			return
 		}
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
+		return
+	}
+
+	if group.Type != "group" {
+		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": "Cannot add members to a private chat"})
 		return
 	}
 
@@ -168,13 +174,18 @@ func RemoveMembers(c *gin.Context) {
 		return
 	}
 
-	var group models.Group
+	var group models.ChatRoom
 	if err := database.Db.Where("id = ?", groupId).First(&group).Error; err != nil {
 		if err == gorm.ErrRecordNotFound {
 			c.JSON(http.StatusNotFound, gin.H{"success": false, "message": "Group not found"})
 			return
 		}
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
+		return
+	}
+
+	if group.Type != "group" {
+		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": "Cannot remove members from a private chat"})
 		return
 	}
 
@@ -223,13 +234,17 @@ func UpdateGroup(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": "Name is required"})
 		return
 	}
-	var group models.Group
+	var group models.ChatRoom
 	if err := database.Db.Where("id = ?", id).First(&group).Error; err != nil {
 		if err == gorm.ErrRecordNotFound {
 			c.JSON(http.StatusNotFound, gin.H{"success": false, "message": "Group not found"})
 			return
 		}
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
+		return
+	}
+	if group.Type != "group" {
+		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": "Cannot update a private chat"})
 		return
 	}
 	group.Name = request.Name
@@ -244,7 +259,7 @@ func UpdateGroup(c *gin.Context) {
 
 func DeleteGroup(c *gin.Context) {
 	id := c.Param("id")
-	var group models.Group
+	var group models.ChatRoom
 	if err := database.Db.Where("id = ?", id).First(&group).Error; err != nil {
 		if err == gorm.ErrRecordNotFound {
 			c.JSON(http.StatusNotFound, gin.H{"success": false, "message": "Group not found"})
@@ -253,9 +268,63 @@ func DeleteGroup(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
 		return
 	}
+	if group.Type != "group" {
+		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": "Cannot delete a private chat"})
+		return
+	}
 	if err := database.Db.Delete(&group).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
 		return
 	}
 	c.JSON(http.StatusOK, gin.H{"success": true, "message": "Group deleted successfully"})
+}
+
+func CreatePrivateChat(c *gin.Context) {
+	var request dto.CreateChatDto
+	if err := c.ShouldBindJSON(&request); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": err.Error()})
+		return
+	}
+
+	var existingUsers []models.User
+	var nonExistingMembers []string
+
+	for _, member := range request.Members {
+		if member == "" {
+			c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": "Member phone no. is required"})
+			return
+		}
+
+		var user models.User
+		result := database.Db.Where("phone = ?", member).First(&user)
+		if result.Error != nil {
+			if result.Error == gorm.ErrRecordNotFound {
+				nonExistingMembers = append(nonExistingMembers, member)
+			} else {
+				c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": result.Error.Error()})
+				return
+			}
+		} else {
+			existingUsers = append(existingUsers, user)
+		}
+	}
+
+	if len(existingUsers) > 0 {
+		group := models.ChatRoom{
+			Type: 	    "private",
+			Name:        request.Name,
+			Description: request.Description,
+			Members:     existingUsers,
+			CreatedAt:   time.Now(),
+			UpdatedAt:   time.Now(),
+		}
+		if err := database.Db.Create(&group).Error; err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": err.Error()})
+			return
+		}
+	}
+
+	// Handle non-existing members (e.g., send invite links)
+	c.JSON(http.StatusOK, gin.H{"success": true, "message": "Private chat created successfully"})
+
 }
